:file_folder: Rename modules
Old 'loop' is new main. Old 'main' is new 'generator'.
